+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - Basic functionality of getting contributors of a repository
 - Support localization
 - Add `lint` package for linting
+- Add ability to customize UI with a builder